Add a new service module for PHP FPM.
This adds a service for the PHP FastCGI Process Manager which replaces
php_fastcgi in future versions. At this stage it consists of unreasonable
default values in the configuration, but we're going to change this soon.

The service needs PHP with compiled in FPM, and thus is based on MarcWeber's
work in his experimental/php branch, so it might make sense to cherry-pick
and/or modify 3d45a854eb50a2895807e4cc33c548af8c755de4 from his nixpkgs fork
before merging in this commit.