postgresql: Use listen_addresses instead of -i.
This replaces the enableTCPIP option with a new listenAddresses option. The
reason is because listen_addresses is by default set to "localhost", so omitting
-i isn't effective, especially in the more recent versions.

Have a look at the point on -i, which essentially deprecates the use of it (and
that since a long time, even for 8.4):

http://www.postgresql.org/docs/9.2/static/app-postgres.html#APP-POSTGRES-OPTIONS

Nevertheless, listen_addresses is the correct way to specify the addresses the
server should listen on. So let's use it and provide an option for users to
change it as necessary.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>