postgresql: Allow to override pg_hba settings.
Until this commit, it was only possible to append to pg_hba.conf, so if you want
to specify mappings and other options, this wasn't possible.

Still, it's only possible to append, but we now restrict the basic configuration
to an absolute minimum, only allowing the root user to connect. That way if you
want to specify access rules for other users, existing auth rules won't get into
the way.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>